+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## NEXT
+
+- Changed the behavior of `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.extend_graph` when
+  specifying a `backward_steps`. Now, nodes and edges will be added as far back that all nodes at `backward_steps` in
+  the past can be inferred. That is, all nodes up to `backward_steps` will now have all their inbound edges and their
+  parent nodes added. This means that the extended graph may now have nodes at lags further back than `backward_steps`.
+
 ## 0.4.5
 
 - Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_adjacency_matrices` for